+++ custom_components/wemportal/number.py
@@ -10,7 +10,7 @@ from homeassistant.helpers.update_coordinator import CoordinatorEntity
 from . import get_wemportal_unique_id
 from homeassistant.helpers.device_registry import DeviceInfo
 from .const import _LOGGER, DOMAIN
-from .utils import (fix_value_and_uom, uom_to_device_class, build_device_info)
+from .utils import (device_model, fix_value_and_uom, uom_to_device_class, build_device_info)
 
 
 async def async_setup_entry(
@@ -176,7 +176,10 @@ class WemPortalNumber(CoordinatorEntity, NumberEntity):
     @property
     def device_info(self) -> DeviceInfo:
         """Get device information."""
-        return build_device_info(self._config_entry.entry_id, self._device_id)
+        return build_device_info(
+            self._config_entry.entry_id, self._device_id,
+            model=device_model(self.coordinator.api, self._device_id),
+        )
 
     @property
     def available(self):